Macros documented (#77), split tests even more
