+++ tests/value.rs
@@ -1,11 +1,7 @@
 #[macro_use]
 extern crate json;
 
-use std::f64;
-use std::collections::BTreeMap;
-use std::collections::HashMap;
-use json::number::Number;
-use json::{ stringify, stringify_pretty, parse, JsonValue, JsonError, Null };
+use json::{ parse, JsonValue, JsonError, Null };
 
 #[test]
 fn is_as_string() {
@@ -121,510 +117,6 @@ fn is_empty() {
     assert!(!json::from(true).is_empty());
     assert!(!array![0].is_empty());
     assert!(!object!{ "foo" => false }.is_empty());
-}
-
-#[test]
-fn stringify_null() {
-    assert_eq!(stringify(Null), "null");
-}
-
-#[test]
-fn stringify_option_none() {
-    let foo: Option<String> = None;
-    assert_eq!(stringify(foo), "null");
-}
-
-#[test]
-fn stringify_option_integer() {
-    let foo = Some(100);
-    assert_eq!(stringify(foo), "100");
-}
-
-#[test]
-fn stringify_str_slice() {
-    assert_eq!(stringify("Foo"), "\"Foo\"");
-}
-
-#[test]
-fn stringify_string() {
-    assert_eq!(stringify("Foo".to_string()), "\"Foo\"");
-}
-
-#[test]
-fn stringify_number() {
-    assert_eq!(stringify(3.141592653589793), "3.141592653589793");
-}
-
-#[test]
-fn stringify_precise_positive_number() {
-    assert_eq!(JsonValue::from(1.2345f64).dump(), "1.2345");
-}
-
-#[test]
-fn stringify_precise_negative_number() {
-    assert_eq!(JsonValue::from(-1.2345f64).dump(), "-1.2345");
-}
-
-#[test]
-fn stringify_zero() {
-    assert_eq!(JsonValue::from(0.0).dump(), "0");
-}
-
-#[test]
-fn stringify_nan() {
-    assert_eq!(JsonValue::from(f64::NAN).dump(), "null");
-}
-
-#[test]
-fn stringify_infinity() {
-    assert_eq!(JsonValue::from(f64::INFINITY).dump(), "null");
-    assert_eq!(JsonValue::from(f64::NEG_INFINITY).dump(), "null");
-}
-
-#[test]
-fn stringify_negative_zero() {
-    assert_eq!(JsonValue::from(-0f64).dump(), "-0");
-}
-
-#[test]
-fn stringify_integer() {
-    assert_eq!(stringify(42), "42");
-}
-
-#[test]
-fn stringify_small_number() {
-    assert_eq!(stringify(0.0001), "0.0001");
-}
-
-#[test]
-fn stringify_large_number() {
-    assert_eq!(stringify(1e19), "10000000000000000000");
-}
-
-#[test]
-fn stringify_very_large_number() {
-    assert_eq!(stringify(3.141592653589793e50), "3.141592653589793e50");
-}
-
-#[test]
-fn stringify_very_large_number_no_fraction() {
-    assert_eq!(stringify(7e70), "7e70");
-}
-
-#[test]
-fn stringify_very_small_number() {
-    assert_eq!(stringify(3.141592653589793e-16), "3.141592653589793e-16");
-}
-
-#[test]
-fn stringify_true() {
-    assert_eq!(stringify(true), "true");
-}
-
-#[test]
-fn stringify_false() {
-    assert_eq!(stringify(false), "false");
-}
-
-#[test]
-fn stringify_array() {
-    assert_eq!(stringify(array![10, false, Null]), "[10,false,null]");
-}
-
-#[test]
-fn stringify_vec() {
-    let mut array: Vec<JsonValue> = Vec::new();
-
-    array.push(10.into());
-    array.push("Foo".into());
-
-    assert_eq!(stringify(array), r#"[10,"Foo"]"#);
-}
-
-#[test]
-fn stringify_typed_vec() {
-    let array = vec![1, 2, 3];
-
-    assert_eq!(stringify(array), "[1,2,3]");
-}
-
-#[test]
-fn stringify_typed_opt_vec() {
-    let array = vec![Some(1), None, Some(2), None, Some(3)];
-
-    assert_eq!(stringify(array), "[1,null,2,null,3]");
-}
-
-#[test]
-fn stringify_object() {
-    let object = object!{
-        "name" => "Maciej",
-        "age" => 30
-    };
-
-    assert_eq!(stringify(object), r#"{"name":"Maciej","age":30}"#);
-}
-
-#[test]
-fn stringify_raw_object() {
-    let mut object = json::object::Object::new();
-
-    object.insert("name", "Maciej".into());
-    object.insert("age", 30.into());
-
-    assert_eq!(stringify(object), r#"{"name":"Maciej","age":30}"#);
-}
-
-#[test]
-fn stringify_btree_map() {
-    let mut map = BTreeMap::new();
-
-    map.insert("name".into(), "Maciej".into());
-    map.insert("age".into(), 30.into());
-
-    // BTreeMap will sort keys
-    assert_eq!(stringify(map), r#"{"age":30,"name":"Maciej"}"#);
-}
-
-#[test]
-fn stringify_hash_map() {
-    let mut map = HashMap::new();
-
-    map.insert("name".into(), "Maciej".into());
-    map.insert("age".into(), 30.into());
-
-    // HashMap does not sort keys, but depending on hashing used the
-    // order can be different. Safe bet is to parse the result and
-    // compare parsed objects.
-    let parsed = parse(&stringify(map)).unwrap();
-
-    assert_eq!(parsed, object!{
-        "name" => "Maciej",
-        "age" => 30
-    });
-}
-
-#[test]
-fn stringify_object_with_put() {
-    let mut object = JsonValue::new_object();
-
-    object["a"] = 100.into();
-    object["b"] = false.into();
-
-    assert_eq!(stringify(object), r#"{"a":100,"b":false}"#);
-}
-
-#[test]
-fn stringify_array_with_push() {
-    let mut array = JsonValue::new_array();
-
-    array.push(100).unwrap();
-    array.push(Null).unwrap();
-    array.push(false).unwrap();
-    array.push(Some("foo".to_string())).unwrap();
-
-    assert_eq!(stringify(array), "[100,null,false,\"foo\"]");
-}
-
-#[test]
-fn stringify_escaped_characters() {
-    assert_eq!(stringify("\r____\n___\t\u{8}\u{c}\\\"__"), r#""\r____\n___\t\b\f\\\"__""#);
-}
-
-#[test]
-fn stringify_dont_escape_forward_slash() {
-    assert_eq!(stringify("foo/bar"), r#""foo/bar""#);
-}
-
-#[test]
-fn stringify_escaped() {
-    assert_eq!(stringify("http://www.google.com/\t"), r#""http://www.google.com/\t""#);
-}
-
-#[test]
-fn stringify_control_escaped() {
-    assert_eq!(stringify("foo\u{1f}bar\u{0}baz"), r#""foo\u001fbar\u0000baz""#);
-}
-
-#[test]
-fn stringify_pretty_object() {
-    let object = object!{
-        "name" => "Urlich",
-        "age" => 50,
-        "parents" => object!{
-            "mother" => "Helga",
-            "father" => "Brutus"
-        },
-        "cars" => array![ "Golf", "Mercedes", "Porsche" ]
-    };
-
-    assert_eq!(stringify_pretty(object, 2),
-               "{\n  \"name\": \"Urlich\",\n  \"age\": 50,\n  \"parents\": {\n    \"mother\": \"Helga\",\n    \"father\": \"Brutus\"\n  },\n  \"cars\": [\n    \"Golf\",\n    \"Mercedes\",\n    \"Porsche\"\n  ]\n}");
-}
-
-#[test]
-fn parse_true() {
-    assert_eq!(parse("true").unwrap(), true);
-}
-
-#[test]
-fn parse_false() {
-    assert_eq!(parse("false").unwrap(), false);
-}
-
-#[test]
-fn parse_null() {
-    assert!(parse("null").unwrap().is_null());
-}
-
-#[test]
-fn parse_number() {
-    assert_eq!(parse("3.141592653589793").unwrap(), 3.141592653589793);
-}
-
-#[test]
-fn parse_small_number() {
-    assert_eq!(parse("0.05").unwrap(), 0.05);
-}
-
-#[test]
-fn parse_very_long_float() {
-    let parsed = parse("2.22507385850720113605740979670913197593481954635164564e-308").unwrap();
-
-    // Handles convertion correctly
-    assert_eq!(parsed.as_f64().unwrap(), 2.225073858507201e-308);
-
-    // Exhausts u64
-    assert_eq!(parsed, Number::from_parts(true, 2225073858507201136, -326));
-}
-
-#[test]
-fn parse_integer() {
-    assert_eq!(parse("42").unwrap(), 42);
-}
-
-#[test]
-fn parse_negative_zero() {
-    assert_eq!(parse("-0").unwrap(), JsonValue::from(-0f64));
-}
-
-#[test]
-fn parse_negative_integer() {
-    assert_eq!(parse("-42").unwrap(), -42);
-}
-
-#[test]
-fn parse_number_with_leading_zero() {
-    assert!(parse("01").is_err());
-}
-
-#[test]
-fn parse_negative_number_with_leading_zero() {
-    assert!(parse("-01").is_err());
-}
-
-#[test]
-fn parse_number_with_e() {
-    assert_eq!(parse("5e2").unwrap(), 500);
-    assert_eq!(parse("5E2").unwrap(), 500);
-}
-
-#[test]
-fn parse_number_with_positive_e() {
-    assert_eq!(parse("5e+2").unwrap(), 500);
-    assert_eq!(parse("5E+2").unwrap(), 500);
-}
-
-#[test]
-fn parse_number_with_negative_e() {
-    assert_eq!(parse("5e-2").unwrap(), 0.05);
-    assert_eq!(parse("5E-2").unwrap(), 0.05);
-}
-
-#[test]
-fn parse_number_with_invalid_e() {
-    assert!(parse("0e").is_err());
-}
-
-#[test]
-fn parse_large_number() {
-    assert_eq!(parse("18446744073709551616").unwrap(), 18446744073709552000f64);
-}
-
-#[test]
-fn parse_array() {
-    assert_eq!(parse(r#"[10, "foo", true, null]"#).unwrap(), array![
-        10,
-        "foo",
-        true,
-        Null
-    ]);
-}
-
-#[test]
-fn parse_object() {
-    assert_eq!(parse(r#"
-
-    {
-        "foo": "bar",
-        "num": 10
-    }
-
-    "#).unwrap(), object!{
-        "foo" => "bar",
-        "num" => 10
-    });
-}
-
-#[test]
-fn parse_object_with_array(){
-    assert_eq!(parse(r#"
-
-    {
-        "foo": [1, 2, 3]
-    }
-
-    "#).unwrap(), object!{
-        "foo" => array![1, 2, 3]
-    });
-}
-
-#[test]
-fn parse_nested_object() {
-    assert_eq!(parse(r#"
-
-    {
-        "l10n": [ {
-            "product": {
-                "inStock": {
-                    "DE": "Lieferung innerhalb von 1-3 Werktagen"
-                }
-            }
-        } ]
-    }
-
-    "#).unwrap(), object!{
-        "l10n" => array![ object!{
-            "product" => object!{
-                "inStock" => object!{
-                    "DE" => "Lieferung innerhalb von 1-3 Werktagen"
-                }
-            }
-        } ]
-    });
-}
-
-#[test]
-fn parse_and_index_from_object() {
-    let data = parse("{ \"pi\": 3.14 }").unwrap();
-    let ref pi = data["pi"];
-
-    assert_eq!(pi, 3.14);
-}
-
-#[test]
-fn parse_and_index_mut_from_object() {
-    let mut data = parse(r#"
-
-    {
-        "foo": 100
-    }
-
-    "#).unwrap();
-
-    assert_eq!(data["foo"], 100);
-
-    data["foo"] = 200.into();
-
-    assert_eq!(data["foo"], 200);
-}
-
-#[test]
-fn parse_and_index_mut_from_null() {
-    let mut data = parse("null").unwrap();
-
-    assert!(data["foo"]["bar"].is_null());
-
-    // test that data didn't coerece to object
-    assert!(data.is_null());
-
-    data["foo"]["bar"] = 100.into();
-
-    assert!(data.is_object());
-    assert_eq!(data["foo"]["bar"], 100);
-
-    assert_eq!(data.dump(), r#"{"foo":{"bar":100}}"#);
-}
-
-#[test]
-fn parse_and_index_from_array() {
-    let data = parse(r#"[100, 200, false, null, "foo"]"#).unwrap();
-
-    assert_eq!(data[0], Number::from(100));
-    assert_eq!(data[1], 200);
-    assert_eq!(data[2], false);
-    assert_eq!(data[3], Null);
-    assert_eq!(data[4], "foo");
-    assert_eq!(data[5], Null);
-}
-
-#[test]
-fn parse_and_index_mut_from_array() {
-    let mut data = parse(r#"[100, 200, false, null, "foo"]"#).unwrap();
-
-    assert!(data[3].is_null());
-    assert!(data[5].is_null());
-
-    data[3] = "modified".into();
-    data[5] = "implicid push".into();
-
-    assert_eq!(data[3], "modified");
-    assert_eq!(data[5], "implicid push");
-}
-
-#[test]
-fn parse_escaped_characters() {
-    let data = parse(r#"
-
-    "\r\n\t\b\f\\\/\""
-
-    "#).unwrap();
-
-    assert!(data.is_string());
-    assert_eq!(data, "\r\n\t\u{8}\u{c}\\/\"");
-}
-
-#[test]
-fn parse_escaped_unicode() {
-    let data = parse(r#"
-
-    "\u2764\ufe0f"
-
-    "#).unwrap();
-
-    assert_eq!(data, "❤️");
-}
-
-#[test]
-fn parse_escaped_unicode_surrogate() {
-    let data = parse(r#"
-
-    "\uD834\uDD1E"
-
-    "#).unwrap();
-
-    assert_eq!(data, "𝄞");
-}
-
-#[test]
-fn parse_escaped_unicode_surrogate_fail() {
-    let err = parse(r#"
-
-    "\uD834 \uDD1E"
-
-    "#);
-
-    assert!(err.is_err());
 }
 
 #[test]
